+++ packages/framer-motion/src/render/dom/utils/is-css-variable.ts
@@ -1,6 +1,8 @@
+export type CSSVariable = `--${string}`
+
 /**
  * Returns true if the provided key is a CSS variable
  */
-export function isCSSVariable(key: string) {
-    return key.startsWith("--")
+export function isCSSVariable(key?: string): key is CSSVariable {
+    return typeof key === "string" && key.startsWith("--")
 }
