+++ packages/framer-motion/src/render/utils/variant-props.ts
@@ -1,0 +1,13 @@
+import { AnimationType } from "./types"
+
+export const variantPriorityOrder = [
+    AnimationType.Animate,
+    AnimationType.InView,
+    AnimationType.Focus,
+    AnimationType.Hover,
+    AnimationType.Tap,
+    AnimationType.Drag,
+    AnimationType.Exit,
+]
+
+export const variantProps = ["initial", ...variantPriorityOrder]
